refactor(vscode-plugin): extract interfaces into types/teamtype.d.ts
